requirements.txt
Flask
openai
python-dotenv
sounddevice
whisper
numpy
scipy